added tricky test option to parametrize (#36)
Test cases can be specified in @parametrize
as tricky and will be excluded from CI (but can be run with pytest -m
"tricky" tests/ so that we can work on and fix them).

Intention is to incentivize writing tests for things that may be tricky
to fix, but without breaking the CI for pull requests etc. E.g. cases
with type promotion, vmap, that may be different in pytorch and jax, or
cases with numerical issues that are tricky to resolve.

Usage examples:

`@parametrize(['jax','torch']).excluding['torch']`

```
@parametrize(['torch', 'jax'], ['float32', 'float64'], op_names).excluding['torch', :, ['psd_identity','psd_scalarmul']]
```

```
@parametrize(['torch', 'jax'], ['float32', 'float64'], op_names).excluding[[('jax','float64','square_kron'), ('torch', 'float32', 'psd_scalarmul')]]
```