Improvements on Sparse API. (#98)
Now the sparse API requires `row_indices`, `col_indices` and `data`
which makes it easier for the user to construct the `Sparse` operator.
This is the COO "ijv" format. Since this format is not the most
efficient for matrix-multiplication, it is then internally transformed
to the CSR format which is what the API was asking the user before.